Conversion to new license, commenting and general tidying up
This commit changes the license to Apache 2.0, adds a bunch of icons to
the tab bar so it looks a little prettier (I've also stuck the SVGs in
there as well). There were a few bugs that I also got rid of so that it
now works on iOS 5. I removed the iPad storyboard. I added a load of
comments in pretty much every file explaining why I did stuff. Pretty
big commit for this projecct!